Bachelors Thesis Daniel Forstenlechner: Adding Unit Tests and cmake support

+++ MATLAB/example_2.m
@@ -4,7 +4,7 @@
 % * Compute a collisionless guiding-center orbit with GORILLA for a passing Deuterium particle.
 % * Use a field-aligned grid for a non-axisymmetric VMEC MHD equilibrium.
 % * Compare the results of GORILLA with different polynominal orders and Runge-Kutta 4.
-% * Create a figure with the Poincaré plots (\varphi = 0) in cylindrical and symmetry flux coordinates.
+% * Create a figure with the Poincarï¿½ plots (\varphi = 0) in cylindrical and symmetry flux coordinates.
 % * Compute the normalized total energy as a function of toroidal mappings. 
 %
 %#######################################################################################################################
@@ -133,19 +133,19 @@ tetra_grid.read();
     %Total Energy of particle in eV
         gorilla_plot.GORILLA_PLOT_NML.energy_eV_start = 3000;
 
-    %Switch for plotting Poincaré cuts at toroidal variable $\varphi$ = 0
+    %Switch for plotting Poincarï¿½ cuts at toroidal variable $\varphi$ = 0
         gorilla_plot.GORILLA_PLOT_NML.boole_poincare_phi_0 = true;
 
-        %Number of skipped (non-printed) Poincaré cuts at toroidal variable $\varphi$ = 0
+        %Number of skipped (non-printed) Poincarï¿½ cuts at toroidal variable $\varphi$ = 0
             gorilla_plot.GORILLA_PLOT_NML.n_skip_phi_0 = 10;
 
-        %Filename for Poincaré cuts at toroidal variable $\varphi$ = 0 in cylindrical coordinates (R,$\varphi$,Z)
+        %Filename for Poincarï¿½ cuts at toroidal variable $\varphi$ = 0 in cylindrical coordinates (R,$\varphi$,Z)
             gorilla_plot.GORILLA_PLOT_NML.filename_poincare_phi_0_rphiz = 'poincare_plot_phi_0_rphiz_order2.dat';
 
-        %Filename for Poincaré cuts at toroidal variable $\varphi$ = 0 in symmetry flux coordinates (s,$\vartheta$,$\varphi$)
+        %Filename for Poincarï¿½ cuts at toroidal variable $\varphi$ = 0 in symmetry flux coordinates (s,$\vartheta$,$\varphi$)
             gorilla_plot.GORILLA_PLOT_NML.filename_poincare_phi_0_sthetaphi = 'poincare_plot_phi_0_sthetaphi_order2.dat';
 
-    %Switch for plotting Poincaré cuts at parallel velocity $v_\parallel$ = 0
+    %Switch for plotting Poincarï¿½ cuts at parallel velocity $v_\parallel$ = 0
         gorilla_plot.GORILLA_PLOT_NML.boole_poincare_vpar_0 = false;
 
     %Switch for plotting full orbit
@@ -195,7 +195,14 @@ gorilla_plot.write([path_RUN,'/gorilla_plot.inp']);
 tetra_grid.write([path_RUN,'/tetra_grid.inp']);
 
 %Create softlinks for used files
-! ln -s ../../../test_gorilla_main.x .
+if isfile('../../../test_gorilla_main.x')
+    ! ln -s ../../../test_gorilla_main.x .
+elseif isfile('../../../BUILD/SRC/test_gorilla_main.x')
+    ! ln -s ../../../BUILD/SRC/test_gorilla_main.x .
+else
+    disp('GORILLA not built, exiting the MatLab script')
+    return
+end
 ! ln -s ../../../MHD_EQUILIBRIA .
 
 %Run GORILLA code
